adk-test-agents/crudo/01-prova-python/requirements.txt
############################
# requirements.txt
############################
google-cloud-run
google-cloud-logging
google-cloud-monitoring
# To list projects - google-cloud-resourcemanager_v3
google-cloud-resource-manager
google-api-python-client # Might be needed for certain discovery aspects or formatting
google-auth # For authentication helpers
google-genai
python-dotenv
pyyaml
# For TZ
pytz
# For Monitoring graphs
matplotlib
# ffor thwe app
streamlit
